Move env hash into separate script
Environment hashing shouldn't be part of the config system, so do it
in a separate Bob script. Add a dependency on this file to build.ninja.
